requirements.txt
asgiref==3.11.1
Django==6.0.3
django-filter==25.2
django-phonenumber-field==8.4.0
djangorestframework==3.16.1
Markdown==3.10.2
phonenumbers==9.0.25
pillow==12.1.1
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
